skill_based/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
astroid==3.3.6
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
dill==0.3.9
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.68.1
gym==0.25.0
gym-notices==0.0.8
gymnasium==1.0.0
idna==3.10
isort==5.13.2
Jinja2==3.1.4
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mccabe==0.7.0
minigrid==3.0.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.29.1
psutil==6.1.0
pydantic==2.10.3
pydantic_core==2.27.1
pygame==2.6.1
pylint==3.3.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.17.0
smmap==5.0.1
stable_baselines3==2.4.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
torch-ac==1.4.0
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.19.1
Werkzeug==3.1.3
